+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### CHANGE LOG
 
+### v2.4.4
+> Updating the function of trackSearchQuerySend in utils.js to send GA events with a callback.
+> Updating the URL queries when making a search request to apachesolr search/site search.
+
 ### v2.4.3
 > Hotfix to get line-height properly set for the dropdowns, which get overridden in a few apps.